+++ src/biginteger_vector.cpp
@@ -9,11 +9,20 @@ biginteger_vector::biginteger_vector(cpp11::strings x) : biginteger_vector(x.siz
       cpp11::check_user_interrupt();
     }
 
-    if (x[i] == NA_STRING) {
+    if (x[i] == NA_STRING || x[i].size() == 0) {
       is_na[i] = true;
     } else {
       try {
-        data[i] = biginteger_type(std::string(x[i]));
+        std::string str(x[i]);
+
+        // remove leading zeros (unless hexadecimal)
+        if (str[0] == '0') {
+          if (str.size() >= 2 && str.compare(0, 2, "0x") != 0 && str.compare(0, 2, "0X") != 0) {
+            str.erase(0, str.find_first_not_of('0'));
+          }
+        }
+
+        data[i] = biginteger_type(str);
       } catch (...) {
         is_na[i] = true;
       }